requirements.txt
keyboard==0.13.5
python-docx==1.1.2